Create LICENSE
Initial commit of ADC-Net project.

ADC-Net is a project designed to help automate procedures for Ophthalmology clinics. Thereby helping to increase the accessibility of screening procedures for eye diseases and improve human health. Please feel free to use the code and make your own meaningful contributions to health and science!